+++ src/dialogs/selectcategoriesdialog.cpp
@@ -83,7 +83,7 @@ void SelectCategoriesDialog::createNewCategory( void )
 		int subcategory = categoryDialog->subcategory();
 
 		//check bounds first
-		if ( result.length() > database->maxCategoryNameLength() ) {
+		if ( result.length() > uint(database->maxCategoryNameLength()) ) {
 			KMessageBox::error( this, QString( i18n( "Category name cannot be longer than %1 characters." ) ).arg( database->maxCategoryNameLength() ) );
 			return ;
 		}
