+++ src/widgets/ingredientinputwidget.cpp
@@ -241,14 +241,14 @@ bool IngredientInput::isInputValid()
 
 bool IngredientInput::checkBounds()
 {
-	if ( ingredientBox->currentText().length() > database->maxIngredientNameLength() ) {
+	if ( ingredientBox->currentText().length() > uint(database->maxIngredientNameLength()) ) {
 		KMessageBox::error( this, QString( i18n( "Ingredient name cannot be longer than %1 characters." ) ).arg( database->maxIngredientNameLength() ) );
 		ingredientBox->lineEdit() ->setFocus();
 		ingredientBox->lineEdit() ->selectAll();
 		return false;
 	}
 
-	if ( unitBox->currentText().length() > database->maxUnitNameLength() ) {
+	if ( unitBox->currentText().length() > uint(database->maxUnitNameLength()) ) {
 		KMessageBox::error( this, QString( i18n( "Unit name cannot be longer than %1 characters." ) ).arg( database->maxUnitNameLength() ) );
 		unitBox->lineEdit() ->setFocus();
 		unitBox->lineEdit() ->selectAll();
@@ -257,7 +257,7 @@ bool IngredientInput::checkBounds()
 
 	QStringList prepMethodList = QStringList::split(",",prepMethodBox->currentText());
 	for ( QStringList::const_iterator it = prepMethodList.begin(); it != prepMethodList.end(); ++it ) {
-		if ( (*it).stripWhiteSpace().length() > database->maxPrepMethodNameLength() )
+		if ( (*it).stripWhiteSpace().length() > uint(database->maxPrepMethodNameLength()) )
 		{
 			KMessageBox::error( this, QString( i18n( "Preparation method cannot be longer than %1 characters." ) ).arg( database->maxPrepMethodNameLength() ) );
 			prepMethodBox->lineEdit() ->setFocus();
